(wip) do not test on browsers that fail
The tests fail on Firefox and Edge.

Reason:
`polymer serve` tries to support all browsers by transforming javascript modules to AMD modules. The bad news is that `moment` has a circular dependency which causes the module to not initialize properly (AMD does not support circular dependencies). Firefox and Edge trigger the modules transformation, because they do not support javascript modules.

The good news is that Firefox added support in v67 and Edge is switching to chromium anyway.

When https://github.com/Polymer/tools/pull/3420 is released, we should add back Firefox to the tested browsers.